Edit and view auth configuration

These functions give more control over and visibility into the auth configuration than gads_auth() does. gads_auth_configure() lets the user specify their own:

OAuth app, which is used when obtaining a user token.

API key. If rgoogleads is de-authorized via gads_deauth(), all requests are sent with an API key in lieu of a token. See the vignette How to get your own API credentials for more. If the user does not configure these settings, internal defaults are used. gads_oauth_app() and gads_api_key() retrieve the currently configured OAuth app and API key, respectively.

Usage

gads_auth_configure(app, path, api_key, developer_token)

gads_auth_cache_path()

gads_open_auth_cache_folder()

gads_api_key()

gads_developer_token()

gads_oauth_app()

Arguments

app

OAuth app, in the sense of httr::oauth_app().

path

JSON downloaded from Google Cloud Platform Console, containing a client id (aka key) and secret, in one of the forms supported for the txt argument of jsonlite::fromJSON() (typically, a file path or JSON string).

api_key

API key.

developer_token

Your Google Ads Developer Token.

Value

gads_auth_configure(): An object of R6 class gargle::AuthState, invisibly.

gads_oauth_app(): the current user-configured httr::oauth_app().

gads_api_key(): the current user-configured API key.

See also

Examples

if (FALSE) {
# see and store the current user-configured OAuth app (probaby `NULL`)
(original_app <- gads_oauth_app())

# see and store the current user-configured API key (probaby `NULL`)
(original_api_key <- gads_api_key())

if (require(httr)) {
  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "YOUR_CLIENT_ID_GOES_HERE",
    secret = "YOUR_SECRET_GOES_HERE"
  )
  google_key <- "YOUR_API_KEY"
  gads_auth_configure(app = google_app, api_key = google_key)

  # confirm the changes
  gads_oauth_app()
  gads_api_key()

  # bring your own app via JSON downloaded from Google Developers Console
  # this file has the same structure as the JSON from Google
  gads_auth_configure(path = app_path)

  # confirm the changes
  gads_oauth_app()

  # use own developer token
  gads_auth_configure(developer_token = 'Your developer token')

}

# restore original auth config
gs4_auth_configure(app = original_app, api_key = original_api_key)
}