Returns a forecast in the form of a time series for the Keyword Plan over the next 52 weeks.

Returns a forecast in the form of a time series for the Keyword Plan over the next 52 weeks.

Usage

gads_keyword_plan_forecast_timeseries(
  keyword_plan_id,
  customer_id = getOption("gads.customer.id"),
  login_customer_id = getOption("gads.login.customer.id"),
  verbose = TRUE
)

Arguments

keyword_plan_id

Keyword plan id, you can get list of your keyword plans using gads_get_report with recource keyword_plan

customer_id

Google Ads client customer id, supports a single account id: "xxx-xxx-xxxx" or a vector of ids from the same Google Ads MCC: c("xxx-xxx-xxxx", "xxx-xxx-xxxx")

login_customer_id

Google Ads manager customer id

verbose

Console log output

Value

tibble with keyword plan historical metrics

See also

Examples

if (FALSE) {
# set client id
gads_set_customer_id('xxx-xxx-xxxx')

# set manager id
gads_set_login_customer_id('xxx-xxx-xxxx')

# get list of plan
plan_data <- gads_get_report(
  resource = 'keyword_plan',
  fields = c('keyword_plan.id')
)

# get keyword historical data
historical_plan_data <- gads_keyword_plan_forecast_timeseries(
 keyword_plan_id = plan_data$keyword_plan_id[1]#'
)

}